partition_parser: add virtual recovery partition is none found
On devices like tissot, there is no "recovery" partition, and the
recovery is usually embedded in the boot image, in the initrd.

lk2nd already has a mechanism to create a virtual "lk2nd" partition;
extend this logic to be able to create a "recovery" partition in the
"boot" partition, so that we can boot onto it.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>